Move categories proxy mode test
